Break test classes into separate files and rename.
