+++ python/tests/util.py
@@ -1,21 +1,14 @@
-"""Integration test for kappa clients"""
 from __future__ import absolute_import, print_function, unicode_literals
 from builtins import dict, str
 
 import sys
-import json
-import subprocess
-import random
-import string
-import time
 import uuid
-
-from os import path
-from datetime import datetime
-
 import unittest
 import nose
+from os import path
+
 import kappy
+
 
 KASIM_DIR = path.normpath(
     path.join(path.dirname(path.abspath(__file__)), *([path.pardir]*2))
@@ -23,10 +16,14 @@ KASIM_DIR = path.normpath(
 BIN_DIR = path.join(KASIM_DIR, "bin")
 MODELS_DIR = path.join(KASIM_DIR, "models")
 
+
 def _get_id(name):
     return "%s-%s" % (name, uuid.uuid1())
 
+
 class _KappaClientTest(unittest.TestCase):
+
+    # Universal Tests =========================================================
 
     def test_file_crud(self):
         print("Getting the runtime object...")
@@ -140,84 +137,8 @@ class _KappaClientTest(unittest.TestCase):
             self.assertEqual(351, len(last_status['series']))
 
 
-class RestClientTest(_KappaClientTest):
-    """ Integration test for kappa client"""
-    def __init__(self, *args, **kwargs):
-        """Initalize test by launching kappa server"""
-        self.websim = path.join(BIN_DIR, "WebSim")
-        self.key = self.generate_key()
-        self.port = 6666
-        self.endpoint = "http://127.0.0.1:{0}".format(self.port)
-        return super(RestClientTest, self).__init__(*args, **kwargs)
-
-    def _mark_log_file(self, msg):
-        """A convenience method to mark sections of a continuous log file."""
-        with open('WebSim_test.log', 'a') as logfile:
-            logfile.write("-"*80 + '\n')
-            logfile.write('%s at %s.\n' % (msg, datetime.now()))
-            logfile.write("-"*80 + '\n')
-        return
-
-    def setUp(self):
-        """Set up unit test by launching client
-
-        Note: This is run automatically by nosetests before EACH test method,
-        so there is no object permanence between tests.
-        """
-        print("Starting server")
-        self._mark_log_file('Starting server')
-        subprocess.Popen([
-            self.websim,
-            '--shutdown-key', self.key,
-            '--port', str(self.port),
-            '--level', 'fatal',  # TODO: This is temporary until --log works.
-            '--log', 'WebSim_test.log'
-            ])
-        time.sleep(1)
-        print("Started")
-        return
-
-    def tearDown(self):
-        """Tear down test by shutting down.
-
-        Note: This is run automatically by nosetests afer EACH test method,
-        so there is no object permanence between tests.
-        """
-        runtime = self.getRuntime("test_proj")
-        print("Closing server...")
-        resp = runtime.shutdown(self.key)
-        self._mark_log_file('Stopping server')
-        time.sleep(1)
-        print("Closed", resp)
-        return
-
-    def getRuntime(self, project_id):
-        return kappy.KappaRest(self.endpoint, project_id)
-
-    @classmethod
-    def generate_key(cls):
-        """Generate random key for kappa server. """
-        return ''.join(random.
-                       SystemRandom().
-                       choice(string.ascii_uppercase + string.digits)
-                       for _ in range(100))
-
-    def test_info(self):
-        """Check if the server can return information about the service."""
-        project_id = _get_id("dummy")
-        runtime = self.getRuntime(project_id)
-        info = runtime.get_info()
-        self.assertIsNotNone('environment_projects' in info)
-        self.assertIsNotNone('environment_build' in info)
-
-class StdClientTest(_KappaClientTest):
-    """ Integration test for kappa client"""
-
-    def getRuntime(self,project_id):
-        return kappy.KappaStd(BIN_DIR)
-
-if __name__ == '__main__':
-    module_name = sys.modules[__name__].__file__
-    print("Running nose for package: %s" % module_name)
-
-    result = nose.run(argv=[sys.argv[0], module_name] + sys.argv[1:])
+def run_nose(fname):
+    import nose
+    fpath = os.path.abspath(fname)
+    print("Running nose for package: %s" % fname)
+    return nose.run(argv=[sys.argv[0], fpath] + sys.argv[1:])
